// app/lib/sdk/validations/user/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Social API
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';


export const PostApiUserRegisterBody = zod.object({
  "firstName": zod.string(),
  "lastName": zod.string(),
  "userName": zod.string(),
  "userGender": zod.string(),
  "email": zod.string(),
  "password": zod.string(),
  "birthDate": zod.string().datetime({"offset":true}),
  "bio": zod.string()
})

export const PostApiUserRegisterResponse = zod.unknown()

export const PostApiUserSignInBody = zod.object({
  "email": zod.string(),
  "password": zod.string()
})

export const PostApiUserSignInResponse = zod.unknown()

export const PostApiUserRefreshTokenBody = zod.object({
  "refreshToken": zod.string()
})

export const PostApiUserRefreshTokenResponse = zod.unknown()

export const PostApiUserForgetPasswordBody = zod.object({
  "email": zod.string()
})

export const PostApiUserForgetPasswordResponse = zod.unknown()

export const postApiUserResetPasswordBodyPasswordMin = 8;
export const postApiUserResetPasswordBodyPasswordMax = 100;


export const postApiUserResetPasswordBodyPasswordRegExp = new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$');


export const PostApiUserResetPasswordBody = zod.object({
  "email": zod.string(),
  "token": zod.string(),
  "password": zod.string().min(postApiUserResetPasswordBodyPasswordMin).max(postApiUserResetPasswordBodyPasswordMax).regex(postApiUserResetPasswordBodyPasswordRegExp),
  "confirmPassword": zod.string()
})

export const PostApiUserResetPasswordResponse = zod.unknown()

export const PostApiUserChangePasswordBody = zod.object({
  "currentPassword": zod.string(),
  "newPassword": zod.string(),
  "confirmPassword": zod.string()
})

export const PostApiUserChangePasswordResponse = zod.unknown()

export const PostApiUserLogoutResponse = zod.unknown()

export const GetApiUserGetUserResponse = zod.unknown()

export const GetApiUserGetUserUserIdParams = zod.object({
  "userId": zod.string()
})

export const GetApiUserGetUserUserIdResponse = zod.unknown()

export const getApiUserSearchQueryPageDefault = 1;
export const getApiUserSearchQueryLimitDefault = 20;

export const GetApiUserSearchQueryParams = zod.object({
  "q": zod.string().optional(),
  "userId": zod.string().optional(),
  "page": zod.number().int().default(getApiUserSearchQueryPageDefault),
  "limit": zod.number().int().default(getApiUserSearchQueryLimitDefault)
})

export const GetApiUserSearchResponse = zod.unknown()

export const PutApiUserUpdateUserBody = zod.object({
  "id": zod.string().optional(),
  "firstName": zod.string().optional(),
  "lastName": zod.string().optional(),
  "userName": zod.string().optional(),
  "birthDate": zod.string().datetime({"offset":true}).optional(),
  "bio": zod.string().optional(),
  "profileImageUrl": zod.string().optional(),
  "isPrivate": zod.boolean().optional()
})

export const PutApiUserUpdateUserResponse = zod.unknown()

export const DeleteApiUserDeleteUserResponse = zod.unknown()

export const PostApiDashboardUserSignInBody = zod.object({
  "email": zod.string(),
  "password": zod.string()
})

export const PostApiDashboardUserSignInResponse = zod.unknown()

export const PostApiDashboardUserRefreshTokenBody = zod.object({
  "refreshToken": zod.string()
})

export const PostApiDashboardUserRefreshTokenResponse = zod.unknown()